Renamed CartDto to CartResponse
