+++ djv-cart/src/main/java/lt/dejavu/cart/service/CartService.java
@@ -1,18 +1,18 @@
 package lt.dejavu.cart.service;
 
 import lt.dejavu.auth.model.db.Address;
-import lt.dejavu.cart.dto.CartDto;
+import lt.dejavu.cart.model.rest.CartResponse;
 import lt.dejavu.payment.exception.PaymentException;
 import lt.dejavu.payment.model.Card;
 
 public interface CartService {
-    CartDto getCart(long userId);
+    CartResponse getCart(long userId);
 
-    CartDto addToCart(long userId, long productId, int amount);
+    CartResponse addToCart(long userId, long productId, int amount);
 
-    CartDto updateAmount(long userId, long productId, int amount);
+    CartResponse updateAmount(long userId, long productId, int amount);
 
-    CartDto removeProduct(long userId, long productId);
+    CartResponse removeProduct(long userId, long productId);
 
     void checkout(long userId, Card cardInfo, Address shippingAddress) throws PaymentException;
 }
