+++ djv-cart/src/main/java/lt/dejavu/cart/service/CartServiceImpl.java
@@ -4,7 +4,7 @@ import lt.dejavu.auth.exception.UserNotFoundException;
 import lt.dejavu.auth.model.db.Address;
 import lt.dejavu.auth.model.db.User;
 import lt.dejavu.auth.repository.UserRepository;
-import lt.dejavu.cart.dto.CartDto;
+import lt.dejavu.cart.model.rest.CartResponse;
 import lt.dejavu.cart.exception.ProductNotInCartException;
 import lt.dejavu.cart.mapper.CartMapper;
 import lt.dejavu.cart.model.db.Cart;
@@ -51,13 +51,13 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public CartDto getCart(long userId) {
+    public CartResponse getCart(long userId) {
         Cart cart = getCartOrCreate(userId);
         return cartMapper.map(cart);
     }
 
     @Override
-    public CartDto addToCart(long userId, long productId, int amount) {
+    public CartResponse addToCart(long userId, long productId, int amount) {
         Cart cart = getCartOrCreate(userId);
         Product product = productRepository.getProduct(productId);
         if (product == null) {
@@ -75,7 +75,7 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public CartDto updateAmount(long userId, long productId, int amount) {
+    public CartResponse updateAmount(long userId, long productId, int amount) {
         Cart cart = getCartOrCreate(userId);
         Optional<OrderItem> itemOpt = CartUtil.getOrderItemByProductId(cart, productId);
         if (!itemOpt.isPresent()) {
@@ -88,7 +88,7 @@ public class CartServiceImpl implements CartService {
     }
 
     @Override
-    public CartDto removeProduct(long userId, long productId) {
+    public CartResponse removeProduct(long userId, long productId) {
         Cart cart = getCartOrCreate(userId);
         cartRepository.removeItem(cart, productId);
         return getCart(userId);
@@ -97,7 +97,7 @@ public class CartServiceImpl implements CartService {
     @Override
     @Transactional
     public void checkout(long userId, Card cardInfo, Address shippingAddress) throws PaymentException {
-        CartDto cart = getCart(userId);
+        CartResponse cart = getCart(userId);
         Payment payment = buildPayment(cart, cardInfo);
         OrderDto order = buildOrder(cart, shippingAddress);
 
@@ -106,7 +106,7 @@ public class CartServiceImpl implements CartService {
         cartRepository.deleteCart(userId);
     }
 
-    private OrderDto buildOrder(CartDto cart, Address shippingAddress) {
+    private OrderDto buildOrder(CartResponse cart, Address shippingAddress) {
         OrderDto order = new OrderDto();
         order.setCreatedDate(Timestamp.from(Instant.now()));
         order.setUser(cart.getUser());
@@ -117,7 +117,7 @@ public class CartServiceImpl implements CartService {
         return order;
     }
 
-    private Payment buildPayment(CartDto cart, Card card) {
+    private Payment buildPayment(CartResponse cart, Card card) {
         Payment payment = new Payment();
         payment.setCard(card);
         payment.setAmount(priceToCents(cart.getTotal()));
